--- apps/swirl-docs/src/components/ComponentExamples/index.ts
@@ -1,1 +1,0 @@
-export * from "./ComponentExamples";
