+++ apps/swirl-docs/src/components/ComponentPreview/index.ts
@@ -1,0 +1,1 @@
+export * from "./ComponentPreview";
